requirements.txt
packaging==24.2
pillow==11.0.0
PyPDF2==3.0.1
pytesseract==0.3.13
